+++ gan/mapmaker/mapimg.py
